feat: Auto-update dependencies when labelled
# Why?
Following on from #1518 - GitHub have updated how permissions in GitHub actions work.  Unfortunately this means that Dependabot workflows no longer have access to any read/write secrets.

# What?
Update to run in the `pull_request_target` scope, for when a label has been added (or removed).

This PR also updates the condition to check the pull_request payload for its original creator instead of relying on the individual actor - so that it is easier to debug!