fix: Replace % with fr for articles
Why?
Article layout makes use of CSS grid to separate the side-bar and the content well.  In many cases it overflows due to the addition of percentages and em units.

What?
Replace usage of percentages in grid-template-columns definitions with fr, which take into account free space available after the gap (in em).