Create robots.txt
# Why?

I don't currently have one of these, and so bots default to scraping this site.  What has made this more urgent is that GPTbot is now scraping websites to ingest content, and I want to explicitly opt out 

# What?
Add a robots.txt adding a crawl delay of 1 second and blocking GPTbot entirely.